pyproject.toml
[tool.poetry]
name = "poetry-codeartifact-auth-plugin"
version = "0.1.0"
description = "Authenticates against AWS CodeArtifact"
authors = ["andy.mackinlay <andy.mackinlay@cultureamp.com>"]
packages = [
  { include = "poetry_codeartifact_auth.py" },
]

[tool.poetry.dependencies]
python = "^3.8"
poetry = "^1.0"
python-dotenv = "^0.21.0"
boto3 = "^1.24.75"
boto3-stubs = {extras = ["codeartifact"], version = "^1.24.75"}

[tool.poetry.dev-dependencies]

[tool.poetry.group.dev.dependencies]
pytest = "^7.1.3"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.poetry.plugins."poetry.application.plugin"]
placeholder-command = "poetry_codeartifact_auth:CodeArtifactAuthPlugin"